Move protocol_param.hpp to silkworm/chain
