+++ core/silkworm/execution/evm_test.cpp
@@ -17,10 +17,10 @@
 #include "evm.hpp"
 
 #include <catch2/catch.hpp>
+#include <silkworm/chain/protocol_param.hpp>
 #include <silkworm/state/memory_buffer.hpp>
 
 #include "address.hpp"
-#include "protocol_param.hpp"
 
 namespace silkworm {
